Adjust to new Bookalope server API v1.2.0: IDML export.
With the latest Bookalope core and server changes, ICML file export is
being replaced by (proper) IDML file export. As a consequence of that,
the extension for InDesign must load and display IDML files instead of
placing ICML Story files.

Note that the new version of the extension happens to be the same as the
server API version: 1.2.0. That’s purely coincidental though 😉